Split AddComment and ListComment servlets
